+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   * Vehicle sheets have undergone some improvements, with new tabs being introduced
   * You can now directly roll a vehicle weapon using a crew member instead of having to select the crew member first
   * Vehicle crew selection is now a multi-select dropdown instead of requiring you to drag-and-drop the same actor multiple times
+  * Dragging crew onto vehicles now prompts for the initial role selection, instead of assigning "(none)"
+  * A single, built-in piloting skill has been added to vehicles, which automatically selects the proper skill for rolling that type of vehicle
 * Fixes:
   * Sending a signature ability to chat now includes purchased upgrades
   * Granting XP to the entire group now updates the XP logs